+++ common/src/main/java/net/mehvahdjukaar/amendments/common/recipe/CauldronRecipe.java
@@ -1,0 +1,14 @@
+package net.mehvahdjukaar.amendments.common.recipe;
+
+import net.minecraft.core.NonNullList;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.crafting.CraftingBookCategory;
+import net.minecraft.world.item.crafting.Ingredient;
+import net.minecraft.world.item.crafting.ShapelessRecipe;
+
+public class CauldronRecipe extends ShapelessRecipe {
+    public CauldronRecipe(ResourceLocation id, String group, CraftingBookCategory category, ItemStack result, NonNullList<Ingredient> ingredients) {
+        super(id, group, category, result, ingredients);
+    }
+}