+++ common/src/main/java/net/mehvahdjukaar/amendments/common/recipe/DummyContainer.java
@@ -1,24 +1,25 @@
 package net.mehvahdjukaar.amendments.common.recipe;
 
 
+import net.minecraft.core.NonNullList;
 import net.minecraft.util.Mth;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.entity.player.StackedContents;
 import net.minecraft.world.inventory.CraftingContainer;
 import net.minecraft.world.item.ItemStack;
 
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 
 public class DummyContainer implements CraftingContainer {
 
-    private final List<ItemStack> items = new ArrayList<>();
+    private final NonNullList<ItemStack> stacks;
     private final int dimension;
 
     private DummyContainer(Collection<ItemStack> items) {
-        this.items.addAll(items);
         this.dimension = Mth.ceil(Math.sqrt(items.size()));
+        this.stacks = NonNullList.withSize(dimension * dimension, ItemStack.EMPTY);
+        this.stacks.addAll(items);
     }
 
     public static DummyContainer surround(ItemStack dye, ItemStack toRecolor) {
@@ -37,24 +38,25 @@ public class DummyContainer implements CraftingContainer {
     public static DummyContainer of(ItemStack... items) {
         return new DummyContainer(List.of(items));
     }
+
     public static DummyContainer of(Collection<ItemStack> items) {
         return new DummyContainer(items);
     }
 
     @Override
     public int getContainerSize() {
-        return items.size();
+        return stacks.size();
     }
 
     @Override
     public boolean isEmpty() {
-        return items.isEmpty();
+        return stacks.isEmpty();
     }
 
     @Override
     public ItemStack getItem(int slot) {
         if (slot >= this.getContainerSize()) return ItemStack.EMPTY;
-        return items.get(slot);
+        return stacks.get(slot);
     }
 
     @Override
@@ -96,7 +98,7 @@ public class DummyContainer implements CraftingContainer {
 
     @Override
     public List<ItemStack> getItems() {
-        return items;
+        return stacks;
     }
 
     @Override