+++ forge/src/main/java/net/mehvahdjukaar/amendments/forge/Test.java
@@ -1,0 +1,11 @@
+package net.mehvahdjukaar.amendments.forge;
+
+import net.mehvahdjukaar.moonlight.api.client.util.RenderUtil;
+import net.minecraft.client.renderer.block.BlockRenderDispatcher;
+
+public class Test {
+
+    public static  void aa(){
+
+    }
+}